+++ examples/resources/meraki_appliance_traffic_shaping_vpn_exclusions/import.sh
@@ -1,0 +1,1 @@
+terraform import meraki_appliance_traffic_shaping_vpn_exclusions.example "<network_id>"