+++ CHANGELOG.md
@@ -34,6 +34,7 @@
 - Add `meraki_appliance_traffic_shaping_rules` resource and data source
 - Add `meraki_appliance_traffic_shaping_uplink_bandwidth` resource and data source
 - Add `meraki_appliance_traffic_shaping_uplink_selection` resource and data source
+- Add `meraki_appliance_traffic_shaping_vpn_exclusions` resource and data source
 
 ## 0.1.1